Rename pyomo.util to pyomo.common to bring this PR up to date with #502
