+++ pyomo/core/base/plugin.py
@@ -36,9 +36,9 @@ __all__ = ['pyomo_callback',
 
 import logging
 import pyutilib.misc
-from pyomo.util.deprecation import deprecated
-from pyomo.util.modeling import unique_component_name
-from pyomo.util.plugin import (
+from pyomo.common.deprecation import deprecated
+from pyomo.common.modeling import unique_component_name
+from pyomo.common.plugin import (
     alias, implements, Interface, Plugin, PluginFactory, CreatePluginFactory,
     PluginError, ExtensionPoint )
 from pyomo.common.timing import TransformationTimer
